backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
crewai==0.70.1
python-dotenv==1.0.0
google-generativeai==0.3.2
crewai-tools==0.8.3